requirements.txt
djangorestframework==3.15.2
pillow==10.4.0
sqlparse==0.5.1